refactor(absinthe): remove node module
